apps/api/pyproject.toml
[project]
name = "packetpilot-api"
version = "0.1.0"
description = "PacketPilot FastAPI service"
requires-python = ">=3.12"
dependencies = [
  "bcrypt>=4.3.0,<5.0.0",
  "fastapi>=0.116.0,<0.117.0",
  "pyjwt>=2.10.1,<3.0.0",
  "pydantic[email]>=2.12.4,<3.0.0",
  "sqlalchemy>=2.0.44,<3.0.0",
  "uvicorn>=0.35.0,<0.36.0"
]

[dependency-groups]
dev = [
  "black>=25.9.0,<26.0.0",
  "httpx>=0.28.1,<0.29.0",
  "pytest>=8.4.2,<9.0.0",
  "ruff>=0.13.2,<0.14.0"
]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
pythonpath = ["."]
